Extract payload decoders into their own files
