+++ Domain Model/EurofurenceModel/Private/Services/Notifications/Payload Decoding/NotificationPayloadDecoder.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+class NotificationPayloadDecoder {
+    
+    private let nextComponent: NotificationPayloadDecoder?
+    
+    init(nextComponent: NotificationPayloadDecoder?) {
+        self.nextComponent = nextComponent
+    }
+    
+    func process(context: NotificationPayloadDecodingContext) {
+        nextComponent?.process(context: context)
+    }
+    
+}
