+++ Domain Model/EurofurenceModel/Private/Services/Notifications/Payload Decoding/SuccessfulSyncNotificationPayloadDecoder.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+class SuccessfulSyncNotificationPayloadDecoder: NotificationPayloadDecoder {
+    
+    override func process(context: NotificationPayloadDecodingContext) {
+        context.complete(content: .successfulSync)
+    }
+    
+}
